requirements.txt
spotipy==2.4.4
cleverwrap==0.2.3.6
discord==0.0.2
discord.py==1.3.4
praw==6.1.1
prawcore==1.0.1
boto==2.49.0
mysql-connector-python==8.0.15
python-valve==0.2.1
asyncpg==0.20.1